+++ client/src/components/Pages/OrgDashboard/AddCoupons/index.js
@@ -6,16 +6,15 @@ import { extendMoment } from 'moment-range';
 import { Row, Col } from '../../../Common/Grid';
 import Title from '../../../Common/Title';
 import * as T from '../../../Common/Typography';
-import { Select, DatePicker, Input } from '../../../Common/Inputs';
-import Icon from '../../../Common/Icon';
-import Button from '../../../Common/ButtonNew';
+
+import Form from './Form';
 
 import { formatPrice, calculatePrice } from '../../../../helpers';
 import { typographies } from '../styleProperties';
-import { Warning, CancelLink, Error } from './style';
-
-import { API_ACCOUNT_URL } from '../../../../constants/apiRoutes';
-import { DASHBOARD_URL } from '../../../../constants/navRoutes';
+import {
+  API_ACCOUNT_URL,
+  API_COUPONS_URL,
+} from '../../../../constants/apiRoutes';
 import { TABLET_WIDTH } from '../../../../constants/screenWidths';
 
 import validationSchema from './validationSchema';
@@ -48,23 +47,16 @@ const initialState = {
   email: '',
   name: '',
   message: '',
-  discountRate: 0,
+  discountRate: null,
   multiDateRange: [
     {
       startDate: null,
       endDate: null,
     },
   ],
-  couponPrice: 0,
+  couponPrice: null,
   errors: {},
 };
-
-const optionsPercentages = [
-  { label: '10', value: 10 },
-  { label: '25', value: 25 },
-  { label: '50', value: 50 },
-  { label: '100', value: 100 },
-];
 
 const AddCoupons = props => {
   const { windowWidth } = props;
@@ -85,7 +77,7 @@ const AddCoupons = props => {
     multiDateRange,
     discountRate,
     couponPrice,
-    errors,
+    code,
   } = state;
 
   useEffect(() => {
@@ -95,7 +87,7 @@ const AddCoupons = props => {
       const { data } = await axios.get(API_ACCOUNT_URL);
 
       if (mounted) {
-        setBalance(formatPrice(data.currentBalance));
+        setBalance(data.currentBalance);
         setBalanceLoading(false);
       }
     }
@@ -108,13 +100,11 @@ const AddCoupons = props => {
 
   // calculate coupon price
   useEffect(() => {
-    const { startDate, endDate } = multiDateRange[0];
+    const { startDate, endDate } = multiDateRange && multiDateRange[0];
 
-    if (startDate && endDate && discountRate > 0) {
+    if (startDate && endDate && discountRate) {
       const range = moment.range(startDate, endDate);
-      const price = formatPrice(
-        (calculatePrice(range) * Number(discountRate)) / 100,
-      );
+      const price = (calculatePrice(range) * Number(discountRate)) / 100;
 
       dispatch({
         type: 'change',
@@ -166,49 +156,62 @@ const AddCoupons = props => {
     e.preventDefault();
 
     try {
-      // setSubmitting(true);
-      if (balance < couponPrice) {
-        setError('Not enough funds');
-        // setSubmitting(false);
-      }
+      setSubmitting(true);
 
       const { errors: validationErrors } = validate({
         schema: validationSchema,
-        data: { ...state },
+        data: { discountRate, multiDateRange, name, email, message },
       });
 
       if (!validationErrors) {
-        // do call
-        // const { data } = await axios.post(API_COUPONS_URL, {
-        //   name,
-        //   email,
-        //   discountRate,
-        //   startDate: moment(multiDateRange[0].startDate).format('YYYY-MM-DD'),
-        //   endDate: moment(multiDateRange[0].startDate).format('YYYY-MM-DD'),
-        // });
-        // setError('');
-        // dispatch({ type: 'isError', errors: null });
-        console.log('ALL GOOD!', {
+        setError('');
+        dispatch({ type: 'isError', errors: {} });
+
+        const {
+          data: { code: _code },
+        } = await axios.post(API_COUPONS_URL, {
           name,
           email,
           discountRate,
           startDate: moment(multiDateRange[0].startDate).format('YYYY-MM-DD'),
           endDate: moment(multiDateRange[0].startDate).format('YYYY-MM-DD'),
         });
-        // setSubmitting(false);
+
+        if (code) {
+          dispatch({ type: 'change', name: 'code', value: _code });
+        }
+
+        setSubmitting(false);
       } else {
         // set validation errors
         dispatch({ type: 'isError', errors: validationErrors });
-        // setSubmitting(false);
+        setSubmitting(false);
       }
     } catch (err) {
       // set server error
       setError(SERVER_ERROR);
-      // setSubmitting(false);
+      setSubmitting(false);
     }
   };
-  console.log('state', state);
-  console.log('err', errors);
+
+  if (code) {
+    return (
+      <>
+        <Row>
+          <Title withBg caps>
+            <Col w={[4, 12, 12]}>success!</Col>
+          </Title>
+        </Row>
+        <Row>
+          <Col w={[4, 10, 7]}>
+            A discount code has been created and sent to {email}. You can also
+            copy the link below if you would like to send it to them manually.{' '}
+          </Col>
+        </Row>
+      </>
+    );
+  }
+
   return (
     <>
       <Row>
@@ -219,115 +222,22 @@ const AddCoupons = props => {
       <Row>
         <Col w={[4, 12, 12]}>
           <AmountTitle color="pink">
-            £{balanceLoading ? '...' : balance}
+            £{balanceLoading ? '...' : formatPrice(balance)}
           </AmountTitle>
           <T.PSBold color="darkBlue">available Balance</T.PSBold>
         </Col>
       </Row>
-      <form onSubmit={handleSubmit}>
-        <Row mt={5} mb={5}>
-          <Col w={[4, 4, 5]}>
-            <Select
-              options={optionsPercentages}
-              label="Discount %"
-              placeholder="%"
-              onChange={val => handleSelectChange(val)}
-              value={discountRate}
-              error={errors.discountRate && 'required'}
-            />
-          </Col>
-        </Row>
-        <Row mb={5}>
-          <Col w={[4, 12, 8]}>
-            {multiDateRange.map((date, index) => (
-              <DatePicker
-                label="Internship dates"
-                onChange={onRangeChange}
-                type="dateRange"
-                index={index}
-                mb={1}
-                value={multiDateRange[index]}
-                error={errors['multiDateRange[0]'] && 'required'}
-              />
-            ))}
-          </Col>
-        </Row>
-        <Row mb={5}>
-          <Col w={[4, 12, 5]}>
-            <Input
-              onChange={handleInputChange}
-              name="email"
-              label="Intern’s email address"
-              placeholder="Enter email address"
-              type="email"
-              value={email}
-              error={errors.email}
-            />
-          </Col>
-        </Row>
-        <Row mb={5}>
-          <Col w={[4, 12, 5]}>
-            <Input
-              onChange={handleInputChange}
-              name="name"
-              label="Intern’s full name"
-              placeholder="Enter name"
-              type="text"
-              value={name}
-              error={errors.name}
-            />
-          </Col>
-        </Row>
-        <Row mb={5}>
-          <Col w={[4, 3, 3]}>
-            <AmountTitle color="pink">£{couponPrice}</AmountTitle>
-            <T.PSBold color="darkBlue">potential cost of internship</T.PSBold>
-          </Col>
-          {notEnoughFunds && (
-            <Warning w={[4, 3, 3]}>
-              <Icon
-                color="pink"
-                icon="warning"
-                width="40"
-                height="40"
-                margin="0 5px 0 0"
-              />
-              <T.H8C caps color="blue" style={{ display: 'inline-flex' }}>
-                Not enough money in the fund
-              </T.H8C>
-            </Warning>
-          )}
-        </Row>
-        <Row mb={5}>
-          <Col w={[4, 12, 8]}>
-            <Input
-              onChange={handleInputChange}
-              name="message"
-              label="Send an optional message to intern"
-              placeholder="Type here..."
-              textArea
-              value={message}
-              error={errors.message}
-            />
-          </Col>
-        </Row>
-        <Row mb={5}>
-          <Col w={[4, 6, 5.4]} mb={5}>
-            <Button type="secondary" bgColor="pink">
-              CREATE DISCOUNT CODE
-            </Button>
-          </Col>
-          <Col w={[4, 6, 5.4]}>
-            <CancelLink
-              to={DASHBOARD_URL}
-              // disabled={paymentLoading}
-            >
-              Cancel
-            </CancelLink>
-          </Col>
-        </Row>
-        {error && <Error>{error}</Error>}
-      </form>
+      <Form
+        state={state}
+        handleSubmit={handleSubmit}
+        handleSelectChange={handleSelectChange}
+        onRangeChange={onRangeChange}
+        handleInputChange={handleInputChange}
+        notEnoughFunds={notEnoughFunds}
+        submitting={submitting}
+        error={error}
+        AmountTitle={AmountTitle}
+      />
     </>
   );
 };
